[ADD] Add WHOIS, RDAP for .ad domain https://domainincite.com/29820-ad-domains-to-go-global-soon
From 22.10.2024 any natural or legal person will be able to register an available and live .ad domain, through any of the accredited registrars, following the standard domain registration procedure, meaning that requests will be handled automatically and the domains will be assigned in strict order of arrival ( first-come, first-served ).
No priority will need to be proven, and applications will not be subject to prior validation.